Use new mod format
