+++ src/labbie/mods.py
@@ -1,6 +1,6 @@
 import dataclasses
 import functools
-from typing import Dict, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple, Union
 
 import injector
 import loguru
@@ -36,44 +36,53 @@ class Mods:
     def helm_mods(self):
         return sorted(self.helm_mod_info.keys())
 
-    def _build_helm_mod_info(self, mods: Dict[str, Tuple[str, Optional[int]]]) -> Dict[str, HelmModInfo]:
+    def _build_helm_mod_info(self, mods: List[List[Tuple[str, List[str], List[int]]]]) -> Dict[str, HelmModInfo]:
         result = {}
-        for mod_format, (slot_patterns, values) in mods.items():
-            if '{' not in mod_format:
-                trade_text = mod_format.lower()
-                result[mod_format] = HelmModInfo(
-                    mod=mod_format,
+        for index, mod_variants in enumerate(mods):
+            for mod_format, slot_patterns, values in mod_variants:
+                if '{' not in mod_format:
+                    trade_text = mod_format.lower()
+                    if (trade_stat_id := self._trade.text_to_stat_id.get(trade_text)) is None:
+                        continue
+                    result[mod_format] = HelmModInfo(
+                        mod=mod_format,
+                        trade_text=trade_text,
+                        trade_stat_id=trade_stat_id,
+                        trade_stat_value=None
+                    )
+                    break
+
+                # if # ocurrs in mod_format we need to rethink some logic below
+                assert '#' not in mod_format
+
+                slotted = mod_format.format(*slot_patterns).lower()
+                hash_only_slot_patterns = ['#'] * mod_format.count('{')
+                hash_only_slotted = mod_format.format(*hash_only_slot_patterns).lower()
+                if slotted in self._trade.text_to_stat_id:
+                    trade_text = slotted
+                elif hash_only_slotted in self._trade.text_to_stat_id:
+                    trade_text = hash_only_slotted
+                else:
+                    logger.debug(f'Unable to find trade text for {mod_format=} {slotted=} {hash_only_slotted=}')
+                    continue
+
+                if len(values) == 1:
+                    trade_stat_value = values[0]
+                elif len(values) == 2:
+                    trade_stat_value = sum(values) / 2
+                else:
+                    raise ValueError(f'{mod_format=} has an unexpected number of values - {len(values)=} {values=}')
+
+                slot_values = [pattern.replace('#', str(value)) for pattern, value in zip(slot_patterns, values)]
+                mod = mod_format.format(*slot_values)
+                result[mod] = HelmModInfo(
+                    mod=mod,
                     trade_text=trade_text,
                     trade_stat_id=self._trade.text_to_stat_id.get(trade_text),
-                    trade_stat_value=None
+                    trade_stat_value=trade_stat_value
                 )
-                continue
-
-            # if # ocurrs in mod_format we need to rethink some logic below
-            assert '#' not in mod_format
-
-            slotted = mod_format.format(*slot_patterns).lower()
-            if slotted in self._trade.text_to_stat_id:
-                trade_text = slotted
-            elif (hash_only_slotted := mod_format.format('#', '#').lower()) in self._trade.text_to_stat_id:
-                trade_text = hash_only_slotted
+                break
             else:
-                logger.debug(f'Unable to find trade text for {mod_format=} {slotted=} {hash_only_slotted=}')
-                trade_text = None
+                logger.error(f'Unable to find trade stat info for {index=} {mod_variants=}')
 
-            if len(values) == 1:
-                trade_stat_value = values[0]
-            elif len(values) == 2:
-                trade_stat_value = sum(values) / 2
-            else:
-                raise ValueError(f'{mod_format=} has an unexpected number of values - {len(values)=} {values=}')
-
-            slot_values = [pattern.replace('#', str(value)) for pattern, value in zip(slot_patterns, values)]
-            mod = mod_format.format(*slot_values)
-            result[mod] = HelmModInfo(
-                mod=mod,
-                trade_text=trade_text,
-                trade_stat_id=self._trade.text_to_stat_id.get(trade_text),
-                trade_stat_value=trade_stat_value
-            )
         return result
